resume refactor
Squash Aider commits

feat: Update Read the Docs configuration to resolve build issues

update sql_agent

refactor: Improve SQL agent error handling with max iterations and generate method

fix: Add handle_parsing_errors to SQL agent for better error handling

refactor: Update SQL schema to support structured metadata storage

docs: Add usage documentation to sql_agent.py module

rtds

rtds

rtds

fix: Remove unnecessary langchain_core mock modules in docs conf

docs: Update Sphinx configuration with additional mock modules and source file mappings

fix: Add missing mock modules to resolve Sphinx autodoc warnings

docs: Add missing mock modules to resolve Sphinx autodoc warnings

docs: Fix documentation build warnings by adding mock modules and correcting RST formatting

docs: Fix documentation build warnings and duplicate labels

I'll generate a concise commit message for these changes:

chore: Update documentation dependencies and mock modules

fix: Add missing sphinx-copybutton dependency in requirements.txt

docs: Fix Read the Docs configuration and dependencies

rtds

update Read the Docs files

rtds

Delete .coverage

EOD 11/27

docs: Update documentation requirements with explicit package versions

docs: Add title to modules.rst for Sphinx documentation

chore: Update documentation requirements and modules configuration

sql

fix: Update documentation requirements to resolve sphinx extension issues

rtds

rtd

docs: Fix Read the Docs configuration and module documentation

docs: Update RST header formatting in indigobot documentation

fix: Update query_database calls to use test database path in SQL agent tests

fix: Resolve SQLite database locking and table creation issues in unit tests

fix: Update query_database call with db_path parameter in test case

fix: Update test_sql_agent.py to use test database path and store original DB path

docs: Add Read the Docs configuration files for documentation

delete old RtDocs files

fix: Update SQL agent tests to use explicit test database path

This commit addresses several issues in the SQL agent unit tests:
1. Simplified database setup and teardown
2. Ensured each test uses a clean database
3. Added explicit database path parameter to test methods
4. Removed unnecessary global variable modifications
5. Simplified database clearing between tests

The changes ensure that:
- Each test starts with a clean database
- Test database operations use the correct test database path
- Database setup is more straightforward and less error-prone

fix: Update SQL agent functions to support custom database paths

fix: Resolve database initialization issues in SQL agent tests

fix: Improve SQL agent test database initialization and cleanup

docs: Add clean and html targets to Sphinx Makefile

fix: Update Read the Docs configuration and dependencies

fix: Improve test database management and cleanup in SQL agent tests

fix: Resolve SQLite database locking and connection issues in tests

fix: Resolve SQL agent test failures with database state management

refactor: Improve SQL test setup with cleaner database initialization

fix: Resolve database connection and query issues in SQL agent

fix: Refactor SQL agent tests to improve database handling

fix: Replace self.fail() with pytest.fail() in test_main.py

fix: Update import statement to include rag_chain in test_main.py

fix: Resolve test file import paths and remove undefined function

fix: Correct import paths in test files

fix: Update import path for RecursiveUrlLoader in test_custom_loader.py

refactor: Update import paths in test_refine_html.py

fix: Update import paths and query_database function to resolve test errors

feat: Add new dependencies to documentation requirements

chore: Update Read the Docs configuration and dependencies

near finish of refactor

I'll help you resolve the import issue. The change I suggested looks good. Could you confirm the following:

1. Can you show me the content of `src/indigobot/config.py`?
2. Can you show me the content of `tests/test_main.py`?

These will help me understand the full context and ensure we're resolving the import correctly.

Also, I noticed you mentioned there might be other files we should add. Could you list the files that are related to this import and test?

fix: Update import paths and resolve module import issues

test: Replace unittest.main() with pytest.main() for async test support

test: Refactor base_model tests to use pytest and improve test structure

begin rewiting unit tests with Aider

continue refactor
