+++ .gitignore
@@ -7,7 +7,6 @@ docs/build
 project_indigo.egg-info
 .aider*
 .env
-**/html_files
-**/processed_text
-**/temp
-**/extracted_urls
+**/rag_data
+.vscode
+.coverage
