+++ .readthedocs.yaml
@@ -3,12 +3,11 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.12"
+    python: "3.10"
 
 sphinx:
   configuration: docs/source/conf.py
   fail_on_warning: true
-  builder: dirhtml
 
 formats:
   - pdf
