+++ docs/requirements.txt
@@ -1,23 +1,23 @@
-bs4
-langchain
-langchainhub
-langchain_anthropic
-langchain-chroma
-langchain-community
-langchain-experimental
-langchain_google_genai
-langchain_openai
-langgraph
-pymupdf
-pypdf
-pyyaml
-requests
-setuptools
-setuptools-scm
-sphinx>=7.1.2
-sphinx-rtd-theme>=1.3.0
-sphinx-autodoc-typehints>=1.25.2
-sphinx-copybutton>=0.5.2
-myst-parser>=2.0.0
-sphinx-copybutton>=0.5.2
-unidecode
+# Documentation dependencies
+sphinx==7.2.6
+sphinx-rtd-theme==1.3.0
+sphinx-autodoc-typehints==1.25.2
+sphinx-copybutton==0.5.2
+myst-parser==2.0.0
+
+# Core dependencies that need to be mocked
+langchain-core>=0.1.0
+langchain-community>=0.0.10
+langchain>=0.1.0
+anthropic>=0.7.0
+google-generativeai>=0.3.0
+openai>=1.6.0
+chromadb>=0.4.0
+beautifulsoup4>=4.12.0
+numpy>=1.24.0
+pandas>=2.0.0
+sqlalchemy>=2.0.0
+unidecode>=1.3.0
+
+# Mock imports for docs build
+mock==5.1.0
