+++ docs/source/indigobot.rst
@@ -1,0 +1,37 @@
+indigobot package
+=================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   indigobot.utils
+
+Submodules
+----------
+
+indigobot.config module
+-----------------------
+
+.. automodule:: indigobot.config
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+indigobot.main module
+---------------------
+
+.. automodule:: indigobot.__main__
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: indigobot
+   :members:
+   :undoc-members:
+   :show-inheritance:
