+++ docs/source/indigobot.utils.rst
@@ -1,0 +1,53 @@
+indigobot.utils package
+=======================
+
+Utils Submodules
+--------------------
+
+indigobot.utils.custom\_loader module
+-------------------------------------
+
+.. automodule:: indigobot.utils.custom_loader
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+indigobot.utils.doctool module
+------------------------------
+
+.. automodule:: indigobot.utils.doctool
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+indigobot.utils.jf\_crawler module
+----------------------------------
+
+.. automodule:: indigobot.utils.jf_crawler
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+indigobot.utils.refine\_html module
+-----------------------------------
+
+.. automodule:: indigobot.utils.refine_html
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+indigobot.utils.sql\_agent module
+---------------------------------
+
+.. automodule:: indigobot.utils.sql_agent
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Utils Module Contents
+------------------------
+
+.. automodule:: indigobot.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
