--- docs/source/langchain_app.CCC_scraper.crawler.rst
@@ -1,7 +1,0 @@
-langchain_app.CCC_scraper.crawler
-=================================
-
-.. automodule:: langchain_app.CCC_scraper.crawler
-   :members:
-   :undoc-members:
-   :show-inheritance:
