--- docs/source/langchain_app.CCC_scraper.refine_html.rst
@@ -1,7 +1,0 @@
-langchain_app.CCC_scraper.refine_html
-=====================================
-
-.. automodule:: langchain_app.CCC_scraper.refine_html
-   :members:
-   :undoc-members:
-   :show-inheritance:
