--- docs/source/langchain_app.base_model.rst
@@ -1,7 +1,0 @@
-langchain_app.base_model
-========================
-
-.. automodule:: langchain_app.base_model
-   :members:
-   :undoc-members:
-   :show-inheritance:
