--- docs/source/langchain_app.custom_loader.rst
@@ -1,7 +1,0 @@
-langchain_app.custom_loader
-===========================
-
-.. automodule:: langchain_app.custom_loader
-   :members:
-   :undoc-members:
-   :show-inheritance:
