--- docs/source/langchain_app.sql_agent.rst
@@ -1,7 +1,0 @@
-langchain_app.sql_agent
-=======================
-
-.. automodule:: langchain_app.sql_agent
-   :members:
-   :undoc-members:
-   :show-inheritance:
