+++ docs/source/modules.rst
@@ -1,12 +1,8 @@
-langchain_app
-=============
+Project Indigo Modules
+======================
 
 .. toctree::
    :maxdepth: 4
 
-   langchain_app.base_model
-   langchain_app.custom_loader
-   langchain_app.sql_agent
-   langchain_app.CCC_scraper.crawler
-   langchain_app.CCC_scraper.refine_html
+   indigobot
