+++ requirements.txt
@@ -2,6 +2,7 @@ black
 bs4
 build
 isort
+jq
 langchain
 langchainhub
 langchain_anthropic
@@ -15,6 +16,7 @@ pylama[all]
 pylama[toml]
 pymupdf
 pypdf
+pytest
 pyyaml
 requests
 setuptools
