+++ tests/requirements-test.txt
@@ -2,9 +2,11 @@
 pytest==8.0.0
 pytest-cov==4.1.0
 pytest-mock==3.12.0
+pytest-asyncio==0.23.5
 
 # Application dependencies
 beautifulsoup4>=4.12.2
+jq
 langchain>=0.1.0
 langchainhub>=0.1.14
 langchain-anthropic>=0.1.1
