+++ tests/test_crawler.py
@@ -1,16 +1,17 @@
 import os
 import unittest
-from unittest.mock import Mock, patch
 import xml.etree.ElementTree as ET
-from langchain_app.CCC_scraper.crawler import (
-    start_session,
-    fetch_xml,
-    extract_xml,
-    load_urls,
+from unittest.mock import Mock, patch
+
+from indigobot.utils.jf_crawler import (
     download_and_save_html,
-    parse_url_and_save,
-    parse_url
+    extract_xml,
+    fetch_xml,
+    load_urls,
+    parse_url,
+    start_session,
 )
+
 
 class TestCrawler(unittest.TestCase):
     def setUp(self):
@@ -23,13 +24,13 @@ class TestCrawler(unittest.TestCase):
                 <loc>https://example.com/page2</loc>
             </url>
         </urlset>"""
-        
+
     def test_start_session(self):
         session = start_session()
         self.assertIsNotNone(session)
-        self.assertEqual(session.adapters['https://'].max_retries.total, 5)
+        self.assertEqual(session.adapters["https://"].max_retries.total, 5)
 
-    @patch('requests.Session')
+    @patch("requests.Session")
     def test_fetch_xml_success(self, mock_session):
         mock_response = Mock()
         mock_response.status_code = 200
@@ -40,7 +41,7 @@ class TestCrawler(unittest.TestCase):
         result = fetch_xml("https://example.com", session)
         self.assertEqual(result, b"<xml>test</xml>")
 
-    @patch('requests.Session')
+    @patch("requests.Session")
     def test_fetch_xml_failure(self, mock_session):
         mock_response = Mock()
         mock_response.status_code = 404
@@ -71,7 +72,7 @@ class TestCrawler(unittest.TestCase):
         os.remove("test_urls/test.txt")
         os.rmdir("test_urls")
 
-    @patch('requests.Session')
+    @patch("requests.Session")
     def test_download_and_save_html(self, mock_session):
         mock_response = Mock()
         mock_response.status_code = 200
@@ -80,21 +81,22 @@ class TestCrawler(unittest.TestCase):
 
         test_urls = ["https://example.com/page1"]
         session = mock_session()
-        
-        with patch('os.makedirs'), patch('builtins.open', unittest.mock.mock_open()):
+
+        with patch("os.makedirs"), patch("builtins.open", unittest.mock.mock_open()):
             download_and_save_html(test_urls, session)
 
-    @patch('langchain_app.CCC_scraper.crawler.fetch_xml')
-    @patch('langchain_app.CCC_scraper.crawler.extract_xml')
+    @patch("indigobot.utils.jf_crawler.fetch_xml")
+    @patch("indigobot.utils.jf_crawler.extract_xml")
     def test_parse_url(self, mock_extract_xml, mock_fetch_xml):
         mock_extract_xml.return_value = ["https://example.com/page1"]
         mock_fetch_xml.return_value = self.test_xml.encode()
 
         session = Mock()
         urls = parse_url("https://example.com/sitemap.xml", session)
-        
+
         self.assertEqual(len(urls), 1)
         self.assertEqual(urls[0], "https://example.com/page1")
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     unittest.main()
