+++ tests/test_custom_loader.py
@@ -1,24 +1,14 @@
-import os
-from pathlib import Path
 import unittest
-from unittest.mock import patch, MagicMock
-from bs4 import BeautifulSoup
-from langchain_app.custom_loader import (
-    clean_text,
-    clean_documents,
-    chunking,
-    extract_text,
-    scrape_main,
-    scrape_articles,
-    PDF_PATH,
-)
+from unittest.mock import MagicMock, patch
 
-# PARENT_DIR = os.path.dirname("..")
-# PDF_PATH = Path(
-#     os.path.join(
-#         PARENT_DIR, "langchain_app/rag_data/pdfs/NavigatingLLMsBegginersGuide.pdf"
-#     )
-# )
+from indigobot.utils.custom_loader import (
+    chunking,
+    clean_documents,
+    clean_text,
+    extract_text,
+    scrape_articles,
+    scrape_main,
+)
 
 
 class TestCustomLoader(unittest.TestCase):
@@ -47,7 +37,7 @@ class TestCustomLoader(unittest.TestCase):
         self.assertEqual(cleaned_docs[0].page_content, "Hello World")
         self.assertEqual(cleaned_docs[1].page_content, "Cafe")
 
-    @patch("langchain_app.custom_loader.RecursiveCharacterTextSplitter")
+    @patch("indigobot.utils.custom_loader.RecursiveCharacterTextSplitter")
     def test_chunking(self, mock_splitter):
         """Test chunking function"""
         mock_splitter_instance = MagicMock()
@@ -89,7 +79,7 @@ class TestCustomLoader(unittest.TestCase):
         result = extract_text(html_content_no_main)
         self.assertEqual(result, "Other content")
 
-    @patch("langchain_app.custom_loader.RecursiveUrlLoader")
+    @patch("indigobot.utils.custom_loader.RecursiveUrlLoader")
     def test_scrape_main(self, mock_loader):
         """Test scrape_main function"""
         mock_loader_instance = MagicMock()
@@ -105,8 +95,8 @@ class TestCustomLoader(unittest.TestCase):
         mock_loader_instance.load.assert_called_once()
         self.assertEqual(result, mock_docs)
 
-    @patch("langchain_app.custom_loader.AsyncHtmlLoader")
-    @patch("langchain_app.custom_loader.BeautifulSoupTransformer")
+    @patch("indigobot.utils.custom_loader.AsyncHtmlLoader")
+    @patch("indigobot.utils.custom_loader.BeautifulSoupTransformer")
     def test_scrape_articles(self, mock_transformer, mock_loader):
         """Test scrape_articles function"""
         mock_loader_instance = MagicMock()
@@ -131,14 +121,6 @@ class TestCustomLoader(unittest.TestCase):
             documents=mock_docs, tags_to_extract=[]
         )
         self.assertEqual(result, mock_transformed_docs)
-
-    def test_pdf_path_exists(self):
-        """Test that the PDF file path is valid"""
-        self.assertTrue(os.path.exists(PDF_PATH), f"PDF file not found at {PDF_PATH}")
-        self.assertTrue(
-            str(PDF_PATH).find("/langchain_app/rag_data/"),
-            "PDF path should be relative to langchain_app directory",
-        )
 
 
 if __name__ == "__main__":
